# contract/openapi.v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "testhost | v1",
    "description": "",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost/"
    }
  ],
  "paths": {
...
    "/api/v1/t/{tenantSlug}/claims/{characterId}": {
      "get": {
        "tags": [
          "CharacterClaims"
        ],
        "parameters": [
          {
            "name": "characterId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "tenantSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterClaimServiceModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterClaimServiceModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterClaimServiceModel"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "CharacterClaims"
        ],
        "parameters": [
          {
            "name": "characterId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "tenantSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/t/{tenantSlug}/claims/{characterId}/holder": {
      "delete": {
        "tags": [
          "CharacterClaims"
        ],
        "parameters": [
          {
            "name": "characterId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "tenantSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/characters/{realm}/{name}": {
      "get": {
        "tags": [
          "Characters"
        ],
        "parameters": [
          {
            "name": "realm",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "region",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterDetailServiceModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterDetailServiceModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterDetailServiceModel"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/t/{tenantSlug}/sync/characters/{realm}/{name}": {
      "post": {
        "tags": [
          "TenantSync"
        ],
        "parameters": [
          {
            "name": "realm",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "region",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tenantSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterDetailServiceModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterDetailServiceModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterDetailServiceModel"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CharacterClaimServiceModel": {
        "type": "object",
        "properties": {
          "characterId": {
            "type": "string",
            "format": "uuid"
          },
          "claimedByUserId": {
            "type": [
              "null",
              "string"
            ]
          },
          "claimedByDisplayName": {
            "type": [
              "null",
              "string"
            ]
          },
          "claimedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
      "CharacterClass": {
        "enum": [
          "Warrior",
          "Paladin",
          "Hunter",
          "Rogue",
          "Priest",
          "DeathKnight",
          "Shaman",
          "Mage",
          "Warlock",
          "Monk",
          "Druid",
          "DemonHunter",
          "Evoker"
        ]
      },
      "CharacterDetailServiceModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "realmSlug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "level": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "class": {
            "$ref": "#/components/schemas/CharacterClass"
          },
          "spec": {
            "type": [
              "null",
              "string"
            ]
          },
          "itemLevel": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "faction": {
            "$ref": "#/components/schemas/CharacterFaction"
          },
          "lastSyncedAt": {
            "type": "string",
            "format": "date-time"
          },
          "equipment": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EquippedItemServiceModel"
            }
          },
          "classColor": {
            "type": "string"
          },
          "isDegraded": {
            "type": "boolean"
          }
        }
      },
      "CharacterFaction": {
        "enum": [
          "Alliance",
          "Horde"
        ]
      },
...
      "EquipmentSlot": {
        "enum": [
          "Head",
          "Neck",
          "Shoulder",
          "Back",
          "Chest",
          "Wrist",
          "Hands",
          "Waist",
          "Legs",
          "Feet",
          "Finger1",
          "Finger2",
          "Trinket1",
          "Trinket2",
          "MainHand",
          "OffHand"
        ]
      },
      "EquippedItemServiceModel": {
        "type": "object",
        "properties": {
          "slot": {
            "$ref": "#/components/schemas/EquipmentSlot"
          },
          "blizzardItemId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "itemName": {
            "type": "string"
          },
          "quality": {
            "$ref": "#/components/schemas/ItemQuality"
          },
          "itemLevel": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "iconUrl": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ItemQuality": {
        "enum": [
          "Poor",
          "Common",
          "Uncommon",
          "Rare",
          "Epic",
          "Legendary",
          "Artifact",
          "Heirloom"
        ]
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
    }
  },
...
}